feat: Add create-feature.js script
This commit adds a new script called create-feature.js. This script allows the user to create a new feature branch based on the 'develop' branch. It checks if a branch code is provided as an argument, fetches the latest changes from 'develop', and creates a new branch with the name 'feature/<branchCode>'. If the branch already exists locally or in the remote repository, an error message is displayed.